requirements.txt
aiohttp==3.12.15
motor==3.7.1
Pyrogram==2.0.106
python-dotenv==1.1.1
TgCrypto==1.2.5
